+++ src/main/java/com/comy_delivery_back/dto/response/CupomResponseDTO.java
@@ -4,6 +4,7 @@ import com.comy_delivery_back.enums.TipoCupom;
 import com.comy_delivery_back.model.Cupom;
 import jakarta.validation.constraints.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record CupomResponseDTO (
@@ -11,9 +12,9 @@ public record CupomResponseDTO (
         String codigoCupom,
         String dsCupom,
         TipoCupom tipoCupom,
-        Double vlDesconto,
-        Double percentualDesconto,
-        Double vlMinimoPedido,
+        BigDecimal vlDesconto,
+        BigDecimal percentualDesconto,
+        BigDecimal vlMinimoPedido,
         LocalDateTime dtValidade,
         Integer qtdUsoMaximo,
         Integer qtdUsado,
